refactor: deduplicate code between browser.ts and index.ts
